% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{amoniaco_libre}
\alias{amoniaco_libre}
\title{Cálculo de Amoníaco Libre (NH3L)}
\usage{
amoniaco_libre(NH4, pH, Temp)
}
\arguments{
\item{NH4}{numeric. Nitrógeno amoniacal (id_parametro = 2090) en unidades de
\verb{mg NH4-N/L} (id_unidad = 32).}

\item{pH}{numeric. Potencial de Hidrógeno (id_parametro = 2018).}

\item{Temp}{numeric. Nitrógeno amoniacal (id_parametro = 2032) en unidades de
\code{ºC} (id_unidad = 18).}
}
\value{

}
\description{
Se calcula NH3 a partir de NH4 según \href{ceqg-rcqe.ccme.ca/en/index.html}{Canadian Water Quality Guidelines for the Protection of Aquatic Life} (\href{http://ceqg-rcqe.ccme.ca/download/en/141}{ver Ammonia})
}
