% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{con_sia}
\alias{con_sia}
\title{Conexión con SIA}
\usage{
con_sia()
}
\value{

}
\description{
Genera una conexión con la base de datos infambiental, usando el usuario y
clave de JMB. Sólo funciona en algunas PCs (ej.: servidor o PC de JMB).
}
