% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_comp_cue}
\alias{g_comp_cue}
\title{Gráfico de comparación entre cuencas}
\usage{
g_comp_cue(.data, nombre_clave_param, id_parametro = NULL, ylab = NULL)
}
\arguments{
\item{.data}{data.frame. Datos del SIA}

\item{nombre_clave_param}{character. Nombre clave del parámetro (ej.: "PT").
Se corresponde con la columna \code{nombre_clave} de \code{\link{sia_parametro}}}

\item{id_parametro}{integer. Id del parámetro. Se corresponde con la columna
homónima de \code{\link{sia_parametro}}}

\item{ylab}{character. Etiqueta opcional para usar en la gráfica}
}
\value{

}
\description{
Gráfico de comparación entre cuencas
}
\examples{
d <- filtrar_datos(datos_sia, id_programa = 10L,
                   rango_fechas = c(2017, 2019),
                   id_parametro = 2098)
g_comp_cue(d, "PT")

e <- filtrar_datos(datos_sia, id_programa = 4L,
                   rango_fechas = c(2017, 2019),
                   id_parametro = 2098)
g_comp_cue(e, "PT")
}
