% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{g_mes_all}
\alias{g_mes_all}
\title{Juntar varias gráficas por mes}
\usage{
g_mes_all(.data, id_parametro, t_eti, ...)
}
\arguments{
\item{.data}{Datos}

\item{id_parametro}{\code{integer}. id de varios parámetros}

\item{t_eti}{\code{tbl_df}. Tabla de etiquetas}

\item{...}{Argumentos derivados a \code}
}
\value{

}
\description{
Juntar varias gráficas por mes
}
\examples{
p <- c(PT=2098, NT=2102, ST=2028, Conduc=2009)
dfo <- data.frame(mes = 3:6, codigo_pto = c("RN1", "RN2", "RN3", "RN5"),
                  id_parametro = rep(p, each = 4),
                  valor = rnorm(16))
g_mes_all(dfo, id_parametro = p)
}
\seealso{
\code{\link{g_mes}}, \code{\link{t_eti_base}}
}
