% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{iet}
\alias{iet}
\title{Índice de Estado Trófico}
\usage{
iet(x)
}
\arguments{
\item{x}{numeric: valores de concentración de fósforo total (id_parametro =
2090), en microgramos por litro (\verb{µg P/L}; id_unidad = 1054).}
}
\value{
Un vector numérico con los valores del IET.
}
\description{
Calcula valores del IET a partir de la concentración de Fósforo Total (PT),
en microgramos por litro.
}
\details{
Cálculo: \deqn{IET = 10 \times (6-\frac{0.42-0.36 \times ln(\overline{PT%
(\mu g/L)})}{ln(2)})-20}{% IET = 10 (6 - (.42 - .36 log(PT \mu g/L)) /
log(2)) - 20}
}
\examples{
iet(25)
iet(rlnorm(10, meanlog = 3, 1))
}
