% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lim}
\alias{lim}
\title{Tabla con rangos de valores normales para parámetros}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 132 rows and 14 columns.
}
\usage{
lim
}
\description{
Tabla con rangos de valores normales para parámetros
}
\keyword{datasets}
